--- app/Interfaces/UserActionButtonComponentInterface.php
@@ -1,20 +1,0 @@
-<?php
-
-namespace App\Interfaces;
-
-use App\Models\User;
-
-interface UserActionButtonComponentInterface
-{
-    public function getText(): string;
-
-    public function getRoute(): string;
-
-    public function getIcon(): string;
-
-    public function getUser(): User;
-
-    public function showText(): bool;
-
-    public function isDropdown(): bool;
-}
