+++ app/View/Components/BlockButton.php
@@ -2,25 +2,8 @@
 
 namespace App\View\Components;
 
-use App\Interfaces\UserActionButtonComponentInterface;
-use App\Models\User;
-use Illuminate\View\Component;
-
-class BlockButton extends Component implements UserActionButtonComponentInterface
+class BlockButton extends AbstractUserActionButton
 {
-    public User $user;
-    public bool $dropdown;
-    public bool $showText;
-
-    public function __construct(User $user, bool $showText = false, bool $dropdown = false) {
-        $this->user     = $user;
-        $this->dropdown = $dropdown;
-        $this->showText = $showText;
-    }
-
-    public function render() {
-        return view('components.user-action-button');
-    }
 
     public function getText(): string {
         return $this->user->isBlockedByAuthUser ? __('user.unblock-tooltip') : __('user.block-tooltip');
@@ -32,17 +15,5 @@ class BlockButton extends Component implements UserActionButtonComponentInterfac
 
     public function getIcon(): string {
         return $this->user->isBlockedByAuthUser ? 'fa-unlock' : 'fa-ban';
-    }
-
-    public function getUser(): User {
-        return $this->user;
-    }
-
-    public function showText(): bool {
-        return $this->dropdown || $this->showText;
-    }
-
-    public function isDropdown(): bool {
-        return $this->dropdown;
     }
 }
